MajVj TmaModelPrimitives: add direct wire frame support
To share vertice data in a reasonable way, add wire frame support
inside TmaModelPrimitive directly.

For data with texture, we do not want to share a vertex that have
opposite edge of texture in X position.

For data without texture, we want to share it so to be able to
morph together.