Tma3DScreen: now createStringTexture understands font names including spaces
Before this patch, font names including white spaces need to be quoted
so that it is recognized correctly.

Note: Specifying fallback fonts in some formats won't work.
For instance, Foo, Bar would be assumed as 'Foo, Bar'. To avoid this
problem, needs to quote at least the first font name as 'Foo', Bar.
But basically, this API does not expect fallback fonts, and each caller
makes sure if the specified font is available before calling it.